Adding working mapper, test and role type converter

+++ drupal-mybatis/trunk/src/main/java/org/gbif/user/mybatis/UserRoleTypeConverter.java
@@ -5,16 +5,16 @@ import org.gbif.mybatis.type.BaseConverter;
 
 import com.google.common.collect.ImmutableMap;
 
-public class UserRoleTypeConverter extends BaseConverter<Integer, UserRole>{
+public class UserRoleTypeConverter extends BaseConverter<Long, UserRole>{
 
   /**
    * @throws IllegalArgumentException if two keys have the same value
    */
   public UserRoleTypeConverter() {
-    super(null, new ImmutableMap.Builder<Integer, UserRole>()
-    .put(2, UserRole.USER)
-    .put(3, UserRole.ADMIN)
-    .put(4, UserRole.EDITOR)
+    super(null, new ImmutableMap.Builder<Long, UserRole>()
+    .put(2l, UserRole.USER)
+    .put(3l, UserRole.ADMIN)
+    .put(4l, UserRole.EDITOR)
     .build());
   };
 }
